+++ CHANGELOG.rst
@@ -4,8 +4,8 @@ ChangeLog
 0.1.3 (2019-01-16)
 ------------------
 
-* Added option to use without installed sqlalchemy
-* Added :func:`~pgcom.commuter.Commuter._execute`
+* Added support for the case if sqlalchemy is missing
+* Added :func:`~pgcom.commuter.Commuter._execute` (`#2 <https://github.com/viktorsapozhok/pgcom/issues/2>`_)
 * Added pending transaction handler to :func:`~pgcom.commuter.Commuter.copy_from`
 * Raised :class:`~pgcom.commuter.exc.ExectionError` when execute command fails
 * Replaced :func:`pandas.to_sql` in :func:`~pgcom.commuter.Commuter.insert` by :func:`psycopg.execute_batch`